move group membership logic to Groups repository
